requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
Django
djangorestframework
drf-extra-fields
django-cors-headers
django-storages[s3]
psycopg[binary]
# ML
torch==2.1.1
ultralytics==8.1.2
# Dev
pytest
pytest-django
isort
black
coverage
factory-boy
Pillow